LoginScene.fxml/LoginScene Java Class.
Links to CreateAccountScene.java
